requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
openpyxl==3.11.0
numpy==1.26.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-generativeai==0.3.0
python-dotenv==1.0.0
streamlit==1.29.0
flask==3.0.0
flask-cors==4.0.0
chromadb==0.4.21
